Seperated views file into a package
