+++ CM2301/learn/views/user.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+from learn.forms import *
+
+
+def login(request):
+    form = AuthenticationForm()
+    return render(request, 'login.html', {'form': form})
