+++ CM2301/learn/views/video.py
@@ -1,24 +1,16 @@
 from django.shortcuts import render
 from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.forms import AuthenticationForm
-from django.contrib.auth import authenticate, login
-from django.http import Http404
-from learn.models import Video, User
+from learn.models import Video
 from learn.forms import *
 
-@login_required
-def home(request):
-    values = {}
-    values['title'] = 'Home'
-    return render(request, 'base_navbar.html', values)
 
-def video_create(request):
+def create(request):
     form = VideoUploadForm()
     videos = Video.objects.all()
     return render(request, 'video_upload.html', {'videos': videos, 'form': form})
 
-def videos(request):
+def all(request):
     videos = Video.objects.all()
     return render(request, 'video_list.html', {'videos': videos})
 
@@ -33,7 +25,7 @@ def video(request, video_id):
     return render(request, 'video_player.html', {'video': video})
             
 
-def video_submit(request, video_id):
+def submit(request, video_id):
     if request.method == 'POST':
         form = VideoUploadForm(data=request.POST, files=request.FILES)
         
@@ -41,7 +33,7 @@ def video_submit(request, video_id):
             form.save()
             return HttpResponse("MOTHER FUCKER IT UPLOADED")
         
-def serve_video(request, video_id):
+def serve(request, video_id):
     video = Video.objects.get(pk=video_id)
     print video.uploaded_video.url
     filename = video.uploaded_video.name.split('/')[-1]
@@ -49,16 +41,4 @@ def serve_video(request, video_id):
     print video.uploaded_video.file.size
     response['Content-length'] = video.uploaded_video.file.size
     
-
     return response
-    
-
-def login(request):
-    form = AuthenticationForm()
-    return render(request, 'login.html', {'form': form})
-
-        
-def lecture_create(request):
-    form = LectureCreateForm()
-    print form.as_p()
-    return render(request, 'lecture_create.html', {'form': form})
